Update to upstream tinywl implementation
